GameService layer created to isolate logic and domain from Controller layer
