+++ mastermind-api/src/main/java/com/vanhackathon/mastermind/domain/Colors.java
@@ -1,8 +1,5 @@
 package com.vanhackathon.mastermind.domain;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * All colors possible.
  * 
@@ -10,7 +7,8 @@ import java.util.List;
  */
 public enum Colors {
 
-	RED("R"), BLUE("B");//, GREEN("G"), YELLOW("Y"), PURPLE("P"), ORANGE("O"), CYAN("C"), MAGENTA("M");
+	RED("R"), BLUE("B");// , GREEN("G"), YELLOW("Y"), PURPLE("P"), ORANGE("O"),
+						// CYAN("C"), MAGENTA("M");
 
 	private String color;
 
@@ -21,15 +19,6 @@ public enum Colors {
 	public String getColor() {
 		return color;
 	}
-	
-	public static String parseColors(List<String> colors) {
-		String result = "";
-		for (String string : colors) {
-			result += string;
-		}
-		
-		return result;
-	}
 
 	public static String getColorValues() {
 		Colors[] values = values();
@@ -38,32 +27,7 @@ public enum Colors {
 			Colors colors = values[i];
 			v += colors.getColor();
 		}
-
 		return v;
-	}
-	
-	public static Colors transform(String firstLetter) {
-		Colors[] values = values();
-		for (int i = 0; i < values.length; i++) {
-			String colorFirstLetter = values[i].getColor();
-			if (colorFirstLetter.equals(firstLetter)) {
-				return values[i];
-			}
-		}
-
-		return null;
-	}
-	
-	public static List<Colors> tranformString(String entireSecret) {
-		List<Colors> result = new ArrayList<>();
-		
-		char[] allColors = entireSecret.toCharArray();
-		for (int i = 0; i < allColors.length; i++) {
-			char c = allColors[i];
-			result.add(Colors.transform(String.valueOf(c)));
-		}
-		
-		return result;
 	}
 
 }
